Refactoring user creation now includes ssh keys and set user password
add checks to make sure ./clean.sh and ./build.sh run in appropriate scenarios

+++ clean.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
-
+if [ -e .vagrantfile ]
+  then
+    echo ".vagrantfile still exists. "
+    echo "First run vagrant destroy if you want to create a new environment."
+    exit
+fi
 rm Vagrantfile
 rm manifests/nodes.pp
